pyproject.toml
[tool.poetry]
name = "ada-back"
version = "0.1.0"
description = "Agente personal diario de novedades técnicas para programadores"
authors = ["Karen Rincón"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"

# Agente y razonamiento
langchain = "^1.0"
langchain-google-genai = "^2.0"
langgraph = "^1.0"

# Envío de correo
simplegmail = "^4.1"

# Web scraping
requests = "^2.32"
httpx = "^0.27"
beautifulsoup4 = "^4.12"
lxml = "^5.2"

# Utilidades
python-dotenv = "^1.0"
jinja2 = "^3.1"
pydantic = "^2.12.5"
langchain-openai = "^1.1.10"
langchain-groq = "^1.1.2"
schedule = "^1.2"
tzdata = "^2025.3"
pytz = "^2025.2"
fastapi = "^0.133.1"
uvicorn = {extras = ["standard"], version = "^0.41.0"}

[tool.poetry.group.dev.dependencies]
black = "^24.0"
mypy = "^1.10"
pytest = "^8.2"
pytest-asyncio = "^0.23"

[tool.black]
line-length = 79
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
